Add  users controller
